[Fix #10569] Fix a false positive for `Style/FetchEnvVar`
Fixes #10569.

This PR fixes a false positive for `Style/FetchEnvVar` when using
the same value of `ENV` as `if` condition in the body.

The `ENV` var as a `if` condition is currently allowed.
So the `ENV` var already used in the condition will be accepted as well.

And the following (maybe) corner case is accepted in this PR.

```ruby
if ENV['foo']
  ENV.delete('foo')
  puts 'some code'
  puts ENV['foo']
  puts 'other code'
end
```

https://github.com/rubocop/rubocop/issues/10569#issuecomment-1106213970