Rename `Gemspec/TestFilesAssignment` to `Gemspec/DeprecatedAttributeAssignment`
Follow up https://github.com/rubocop/rubocop/pull/10105#issuecomment-924588815.

This PR renames `Gemspec/TestFilesAssignment` to `Gemspec/DeprecatedAttributeAssignment`.

The already released `Gemspec/DateAssignment` can be integrated in the future.
However that integration is pending as breaking changes cannot be made in RuboCop 1.x.

Anyway, I think #10634 can be able to integrate into the cop of this name.

There is no changelog as this is an unreleased cop naming change.