Add a setting for "autocorrect" in `Naming/InclusiveLanguage`
In this PR, we have made the following word's suggestions
- auto-correct -> autocorrect
- auto_correct -> autocorrect

Also considered "AutoCorrect" but are hesitant to change the
naming of the "RuboCop::Cop::AutoCorrector" module
because of its large impact.

In addition, the "auto-correct" command,
which is scheduled to be deprecated, is suppressed by
including a comment to disable RuboCop in the section containing it.