Fix Layout/SpaceInsideBlockBraces for blocks with numbered arguments
The `Layout/SpaceInsideBlockBraces` was not being applied for blocks with
numbered arguments as they have AST node type of `numblock` and the cop
only ran for `block` nodes.

At Dext we caught spacing errors during code review and wandered why the
cop didn't run at all. The piece of code was in the lines of

```ruby
Tax.where(condition).partition {_1.by_something}
```

Not running Layout/SpaceInsideBlockBraces for `numblock` nodes was the
cause.