Recover 7x slow running `rubocop`
Recovers #10735 regression.

This PR recovers 7x slower running `rubocop`.

## v1.30.1

Originally it was about 12 seconds:

```console
% git chceckout v1.30.1
% bundle exec rubocop --display-time
Inspecting 1366 files
(snip)

1366 files inspected, no offenses detected
Finished in 12.27936500031501 seconds
```

## master (v1.31.0 / after #10735)

7x slower in my local environment:

```console
% bundle exec rubocop --display-time
Inspecting 1393 files
(snip)

1393 files inspected, no offenses detected
Finished in 88.82363200001419 seconds
```

## This patch

It recovers to the same speed as v1.30.1:

```console
% bundle exec rubocop --display-time
Inspecting 1393 files
(snip)

1393 files inspected, no offenses detected
Finished in 12.322518999688327 seconds
```

The cause has not been clarified yet, but it seems that it was caused by
changing `require_relative` to `autoload` in #10735.
This PR change to `require_relative` to resolve the cause of the delay.