Allow asking to avoid mixing hash shorthand styles in single hashes
For those who want to use hash shorthand style, but don't want to mix
shorthand and explicit styles in the same hash we provide a new option
for `EnforcedShorthandSyntax` called `consistent` that raises an offence
if we've mixed styles.  This cop prefers the shorthand syntax only if the
whole hash can be written that way, but if it can't it will prefer fully
explicit values.

E.g.

    # bad
    {foo: , bar: bar}

    # good
    {foo:, bar:}

    # bad
    {foo: , bar: baz}

    # good
    {foo: foo, bar: baz}