Add `RESTRICT_ON_SEND` to the Cop template generated by the `new_cop` task
The generated Cop template uses `on_send`.
In many cases, `RESTRICT_ON_SEND` is used
when `on_send` is to be used as is.

Also Its also possible to make Cop without using `RESTRICT_ON_SEND`,
but it slows down the processing speed, so we want to add it here.

Also added a cop to check for useless `RESTRICT_ON_SEND`
since `RESTRICT_ON_SEND` is defined by default.