Add new `InternalAffairs/RestrictOnSend` cop
This cop will make sure that `RESTRICT_ON_SEND` is
defined if `on_send` or `after_send` is defined.

This comment https://github.com/rubocop/rubocop-rspec/pull/1265#discussion_r861758342
is also mentioned in and leads to optimize the cop
by `RESTRICT_ON_SEND` whenever possible.

However, this PR shall not be used to resolve to the existing cop.

If we included that much in this PR,
the scale would be enormous,
and we thought it would be better to look at
how to optimize each cop in detail.