Preserve multiline semantics on `Style/SymbolArray` and `Style/WordArray`
To fix https://github.com/rubocop/rubocop/issues/10455.

Note that this commit also changes the offense message like this:

```ruby
\# Before
%w(
^^^ Use `[
  'foo',
  'bar'
]` for an array of words.
  foo
  bar
)

\# After
%w(
^^^ Use `[...]` for an array of words.
  foo
  bar
)
```

This means that the auto-corrected code no longer appears in the offense message.
I think this is a reasonable change because the message to enforce the percent styles (`'Use `%w` or `%W` for an array of words.'`) is already using this format.