[Fix #10026] Fix inherit_mode resolution for parent and default config
The problem was that we deleted the `inherit_mode` entry in the
configuration hash while processing the resolution of inheritance from
base configurations (`inherit_from`, etc.). When the merging of default
configuration and user configuration was done at a later stage, the
`inherit_mode` parameter was gone.

By not deleting `inherit_mode`, we fix the problem. I don't think it
has any unwanted side effects, but `inherit_mode` does show up in the
output from `--show-cops` now. This should be OK.