[Fix #10263] Fix aliasing problem for Enabled in Config
There was an aliasing problem in the `Config` class causing the
`cop_options` local variable to reference the same object between
different calls to `for_cop` even when the calls were to different
`Config` objects with different `@hash` objects.

I have not been able to figure out the root cause. Only that the
phenomenon occurs, and that duplicating the cop options before
setting the `Enabled` property resolves it.