[Fix #10613] Allow parallel inspection for autocorrect
Allow the `-P`/`--parallel` flag together with the autocorrection
flags `--[safe-]auto-correct`/`-a` `--auto-correct-all`/`-A` and
`--fix-layout`. Also make this the default, so that `--no-parallel`
has to be given for a single core run.

The algorithm for normal parallel inspection is followed, so that a
cache warm-up phase is run first. This phase only does caching of
inspection results. It doesn't update any files. Then it does a single
core run that takes advantage of the cached results and does the
corrections. The speed gain from this approach is probably smaller
than writing files in parallel, but it's a pretty small change that
still outputs correction results in the same way as before.

Also fix the documentation for `--parallel`, which did not mention
`--no-parallel`.