Support `AllowComments` option for `Style/EmptyElse`
This PR is support `AllowComments` option for `Style/EmptyElse`.
The default for this option is false.

In the case of complex conditional statements,
we may want to leave `else` empty or nil and
write only the explanation of the `else` case in the comments.

`AllowComments: true` will not treat when `else` as offense
if the `else` with an intentional comment.

### @example AllowComments: false (default)

```ruby
# bad
if condition
  statement
else
  # something comment
  nil
end

# bad
if condition
  statement
else
  # something comment
end
```

### @example AllowComments: true

```ruby
# good
if condition
  statement
else
  # something comment
  nil
end

# good
if condition
  statement
else
  # something comment
end
```