[Fix #10926] Make `Style/SafeNavigation` aware of a redundant nil check
Fixes #10926.

This PR makes `Style/SafeNavigation` aware of a redundant nil check.

IMHO, the context of this patch is a minimal update.
OTOH, a new cop like `Style/RedundantNilCheck` cop may be implemented
if it is not suitable for this `Style/SafeNavigation` cop.
(But I'm not sure if it's the best.)