Add a spec for server mode
## Summary

This PR adds a spec for server mode that applies .rubocop.yml configuration
changes even during server startup.

This is already achieved because the server loads .rubocop.yml configuration
for each analysis request from the client.

## Other Information

The feature mentioned in the following article has been implemented for
configuration change of .rubocop.yml.

> You should keep in mind that you'll need to restart the server manually
> if you change RuboCop's configuration.

https://metaredux.com/posts/2022/07/26/rubocop-serves-much-faster.html

The test this PR adds ensure that behavior.