Do not omit parentheses for calls in assignments in conditional branches
Parentheses are required in assignments that call methods in conditional
branches. Removing them is a syntax error and the
`Style/MethodCallWithArgsParentheses` cop with the style of
`omit_parentheses` has been doing just that.

Here is an example that needs the parentheses:

```ruby
if success = destroy_remote(server)
  delete server
end
```

Removing them is a syntax error:

```ruby
if success = destroy_remote server
  delete server
end
```

Such is the case for any assignment or shorthand assignment in any Ruby
conditional.