[Fix #11003] Fix EmptyConditionalBody clobbering
This PR fixes an error that occurs in `Lint/EmptyConditionalBody` when
an `if` branch is empty but the `else` branch is not.

When `rubocop` detects this situation, it will attempt to autocorrect by
removing the `if` block, all the way up to the `else` keyword. It then
attempts to invert the `else` into `unless` - but `else` is already
missing.

Before autocorrection:
```ruby
class SomeClass
  if condition
  else
    do_something
  end
end
```

After autocorrection:
```ruby
class SomeClass
    do_something
  end
end
```

This will cause `parser` to raise a `Parser::Source::TreeRewriter
detected clobbering` exception.

This change checks for the case of an empty `if` body with a present
`else` body, and ensures that only the `if` block is removed(not
including the `else` keyword). Now `else` is still present and can be
inverted into `unless`.