Add new MatchCaseSensitive config in FileName cop
This adds a new config called `MatchCaseSensitive` in the `FileName` cop.
It allows to have matching names without taking into accout the
sensitivity of the file name and the class name, as many eager loading
methods allow it.

With `MatchCaseSensitive: true`, this is not allowed:
- filename: `us_collection.rb`
- classname: `USCollection`

With `MatchCaseSensitive: false`, this is allowed:
- filename: `us_collection.rb`
- classname: `USCollection`