[Fix #11120] Fix an incorrect autocorrect for `Lint/RedundantRequireStatement`
Fixes #11120.

This PR fixes an incorrect autocorrect for `Lint/RedundantRequireStatement`
when using redundant `require` with modifier form.

For example, in essential it is unknown how many times `do_something` will be
executed when `require 'enumerator' while do_something`.
So it does not autocorrects when modifier form is used to avoid side effects.