[RuboCop] RBS Signature Files
This pull request introduces RBS Signature files to the core of RuboCop.
This gives tools such as Steep to perform type-checking, and allows upstream consumes to use our typing to write compatible code.
The idea is to introduce some type-checking and reduce the potential of bugs.
Only works with the new Ruby 3.1