[Fix #5251] Fix loading of configuration in multi-file edge case
Ingredients for the edge-case:
* .rubocop.yml inherits another config (inherit_from/inherit_gem)
* .other config requires an additional gem (e.g. rubocop-rspec)
* .rubocop.yml configures a cop defined in additional gem (e.g. RSpec/VariableName)
  whose name already exists in rubocop (e.g. Naming/VariableName)

What would have happened before is that during the call of add_missing_namespaces,
the registry would have replaced the configuration of the RSpec/VariableName with
Naming/VariableName, because that's the only cop that it knew at that point.

By changing the call order we make sure that requires of external files are
resolved before letting the registry replace badge names.