Add `--[no-]auto-gen-enforced-style` CLI option
This PR is add `--[no-]auto-gen-enforced-style` CLI option.

As a motivation, for a cop that has EnforcedStyle and other options
as configurable options, the offense may remain
even if `--auto-gen-cofig` is done.

Normally, this may not be a problem, but there are cases where you want to update RuboCop,
and you want to make only updates.
For example, if we do not want to include the newly added cop violations up to the point of fixing them,
we can run `rubocop --auto-gen-config` when updating and slowly resolve them later.

The following is a case where `rubocop --auto-gen-config` does not make all offense TODO:
```ruby
# test.rb

a = {

}
```

Run `bundle exec rubocop --auto-gen-config`.
The following TODO configuration file is then created.

```yaml
# .rubocop_todo.yml

# Offense count: 1
# This cop supports safe autocorrection (--autocorrect).
# Configuration parameters: EnforcedStyleForEmptyBraces.
# SupportedStyles: space, no_space, compact
# SupportedStylesForEmptyBraces: space, no_space
Layout/SpaceInsideHashLiteralBraces:
  EnforcedStyle: space
```

If I run `bundle exec rubocop` again in this state, the error still offense.

```
test.rb:1:6: C: [Correctable] Layout/SpaceInsideHashLiteralBraces: Space inside empty hash literal braces detected.
a = { ...
```