Add new `Style/MinMaxComparison` cop
This PR adds new `Style/MinMaxComparison` cop.

This cop Enforces the use of `max` or `min` instead of comparison for greater or less.
It is unsafe because even if a value has `<` or `>` method, it is not necessarily `Comparable`.

```ruby
# bad
a > b ? a : b
a >= b ? a : b

# good
[a, b].max

# bad
a < b ? a : b
a <= b ? a : b

# good
[a, b].min
```