Distribute Rake Tasks for RuboCop extensions
The motivation for this PR is currently that RuboCop RSpec uses the exact same cut_release task as RuboCop does.
We thought that by distributing these on the RuboCop side, we could abolish double maintenance.

With this change, on the RuboCop extension side, the following code can be added to the Rakefile to use the Rake Task defined on any RuboCop side.

```ruby
load "rubocop/tasks/cut_release.rake"
```