[Fix #11338] Add new `Style/ComparableClamp` cop
## Summary

Fixes #11338.

This PR adds new `Style/ComparableClamp` cop.
It enforces the use of `Comparable#clamp` instead of comparison by minimum and maximum.

```ruby
# bad
[[x, low].max, high].min

# bad
if x < low
  low
elsif high < x
  high
else
  x
end

# good
x.clamp(low, high)
```

This cop supports autocorrection for `if/elsif/else` bad style only.
Because `ArgumentError` occurs if the minimum and maximum of `clamp` arguments are reversed.
When these are variables, it is not possible to determine which is the minimum and maximum:

```ruby
[1, [2, 3].max].min # => 1
1.clamp(3, 1)       # => min argument must be smaller than max argument (ArgumentError)
```

## Additional Information

- `Comparable#clamp` was introduced in Ruby 2.4
- Part of #11338 was implemented by #11341