Fix a false negative for `Style/OperatorMethodCall`
This PR fixes a false negative for `Style/OperatorMethodCall`
when using `a.+ b.something`.

The following is the difference depending on whether there is the dot or not.

It is the same when the argument is not enclosed in parentheses:

```console
% ruby-parse -e "a.+ b.something"
(send
  (send nil :a) :+
  (send
    (send nil :b) :something))

% ruby-parse -e "a + b.something"
(send
  (send nil :a) :+
  (send
    (send nil :b) :something))
```

It is different when the argument is enclosed in parentheses.

```console
% ruby-parse -e "a.+(b).something"
(send
  (send
    (send nil :a) :+
    (send nil :b)) :something)

% ruby-parse -e "a +(b).something"
(send nil :a
  (send
    (send
      (begin
        (send nil :b)) :something) :+@))
```

So the former should be warned by this cop.