[Fix #11303] Make `Metrics/ParameterLists` aware of `Struct.new` and `Data.define` blocks
Fixes #11303.

This PR makes `Metrics/ParameterLists` aware of `Struct.new` and `Data.define` blocks.

The number of these `initialize` arguments depends on the number of arguments of
`Struct.new` and `Data.define`, so it wouldn't make sense as a metrics.

One abandoned design, the number of `initialize` argument will check no more than
the number of `Struct.new` and `Data.define` arguments. However, considering
a special case that the `keyword_init` parameter can be included in the number,
I think there is no much benefit for the complexity of the design.

So this PR chooses an approach that doesn't always check for the number of
`initialize` arguments.