Use concurrency only in Pull Request events
Another approach to https://github.com/rubocop/rubocop/pull/11414 .

The following are included as keys to the group.
`${{ github.head_ref || github.run_id }}`

This allows concurrency to work only with Push to PullRequest,
and cancels all but the most recent push for successive pushes;
for merge commits, in other words push events,
`${{ github.run_id }}` is used,
so The cancellation is not enforced because it is always a unique group.