CI against JRuby 9.4 instead of JRuby 9.3
JRuby 9.4.1.0 is now available on setup-ruby.
https://github.com/ruby/setup-ruby/releases/tag/v1.136.0

JRuby 9.4 is MRI 3.1 compatible and JRuby 9.3 is MRI 2.6 compatible.
JRuby 9.3 is still supported as a RuboCop runtime, but presumably any new JRuby issues
found may be fixed in JRuby 9.4. And MRI 2.6 will be dropped from the runtime
in the not distant future.

So, JRuby 9.3 will not be retained and will be replaced by JRuby 9.4 because CI latency.