Fix octal handling in Regexp char classes ...
regexp_parser v2.7.0 introduced a fix that makes it correctly detect octal escapes (e.g. "\141", which equals "a") inside character classes, e.g. `/[\141]/`.

Previously, these were incorrectly interpreted and chunked, c.f. https://github.com/ammar/regexp_parser/issues/86

The incorrect chunking made your build green. This is because `\032` emitted `\0`, which was checked for specifically in `#backslash_zero?`.

The fix in regexp_parser v2.7.0, however, made this spec red.

The correct output by rubocop would be to detect this case as a redundant use of a char class, because it is.

The same goes for the original example "\0":

```ruby
[/\0/.match?("\0"), /[\0]/.match?("\0")] # => [true, true]
```

For `\1` to `\7`, putting them in a char class is not redundant but actually required, because they would otherwise be treated as backreferences.