Add contribution note for CHANGELOG.md
Occasionally I review direct edit to the CHANGELOG.md:

- https://github.com/rubocop/rubocop/pull/11559#discussion_r1101152654
- https://github.com/rubocop/rubocop/pull/11406#discussion_r1064492125
- https://github.com/rubocop/rubocop/pull/11395#discussion_r1064000369

It might be communicated a little more efficiently if listed in the CHANGELOG.md.
And the triple dash for commenting out is used intentionally.
https://stackoverflow.com/questions/4823468/comments-in-markdown/4829998#4829998

It affects when markdown is viewed on GitHub. This is unnecessary note for
user viewing the CHANGELOG.md on GitHub.

Also add changelog/.gitkeep to prevent dead link after release.