Fix parentheses in value omissions for multiple assignments
I'm on a quest to run the `EnforcedShorthandSyntax: always` style on a
big codebase. One of the latest thins I found was a value omission in
multiple assignments. Not the best idea, but it is a real "production"
code usage. Here is an estimation of the code:

```ruby
foo = foo = do_stuff arg, opt1: opt1,
                          opt2: opt2

something.each { pass }
```

The multiple assignment in the code above tricks the current version of
the parentheses detection code and we don't put the parens when
autocorrecting:

```ruby
foo = foo = do_stuff arg, opt1:,
                          opt2:

something.each { pass }
```

While not a syntax error, Ruby is treating it like:

```ruby
foo = foo = do_stuff arg, opt1:,
                          opt2: something.each { pass }
```

We do need to put the parentheses for end result like:

```ruby
foo = foo = do_stuff(arg, opt1:,
                          opt2:)

something.each { pass }
```