Optimize `Style/WordArray` complex matrix check
One of the checks we do is to check if an otherwise simple array is
a member of a parent array consisting exclusively of arrays, including
at least one with "complex" content.

This means that for each sub-array, we navigate up to the parent, and
check all of its children, resulting in `n²` array complexity checks.

By maintaining a cache of checked parent array nodes per investigation
(i.e. per file), we can reduce this to `n` array complexity checks.

This has a massive impact on performance when investigating files
containing large arrays of arrays. A test is added with a timeout that
does not pass when using the non-caching implementation (it would take
over a minute).

---

In passing, this switches the terminology from `2d_array` to `matrix`,
as `2d_array` is inconvenient as it cannot be used as the start of an
identifier.