[Fix #11596] Make `Style/AccessorGrouping` aware of method call before accessor
Fixes #11596.

This PR makes `Style/AccessorGrouping` aware of method call before accessor.

If there is a method call before the accessor method it is always allowed as
it might be intended like Sorbet .
Note, it respects user's intentions, not just Sorbet. Because (annotation) comments
are already allowed, so (annotation) methods can be generalized as well.