Fix a false positive for `Lint/UselessAccessModifier` when using same access modifier inside and outside the `included` block
This PR is fix a false positive for `Lint/UselessAccessModifier` when using same access modifier inside and outside the `included` block.

If an included block is used within a class, as shown below, I think it should not be a violation even if the same access modifier is used inside and outside the included block.

```ruby
class SomeClass
  included do
    private
    def foo; end
  end
  private
  def bar; end
end
```