Improve `Naming/MethodName` example `AllowedPatterns`
None of the code paths lead to scenarios in which a method name could
possibly include leading or trailing whitespace, so there is no reason
for the documentation or test patterns to include optional whitespace.

Instead, we can show an example of anchoring on both ends, to exactly
match the method name.