Add new `Lint/DuplicateMatchPattern` cop
This PR adds new `Lint/DuplicateMatchPattern` cop.
It checks that there are no repeated patterns used in `in` keywords.

```ruby
# bad
case x
in 'first'
  do_something
in 'first'
  do_something_else
end

# good
case x
in 'first'
  do_something
in 'second'
  do_something_else
end
```

I was wondering whether to extend existing `Lint/DuplicateCaseCondition`,
but it is a different inspection target from the concept named "`case` condition".
So, I implemented it as a new `Lint/DuplicateMatchPattern` for "match pattern".