[Fix #11762] Fix an incorrect autocorrect for `Style/ClassEqualityComparison`
Fixes #11762.

This PR fixes an incorrect autocorrect for `Style/ClassEqualityComparison`
when comparing a variable or return value for equality.

when a variable or return value of a method is used, it cannot be suggested
and safe autocorrected because the type is not known.
So this PR tweaks the cop that doesn't suggest `instance_of` argument at those times
and doesn't autocorrect. Anyway, detection is possible.