[Fix #11765] Fix an error for `Style/MultilineMethodSignature`
Fixes #11765.

This PR fixes an error for `Style/MultilineMethodSignature`
when line break after `def` keyword.

It seems preferable that the following cases can be detected by a new different cop:

```ruby
def
method_name arg; end
```

This is probably not a common line break, the new cop will be able to handle `class`, `module`,
class method definitions, and others as well.

And this `Style/MultilineMethodSignature` cop is disabled by default,
but I have a feeling that the new cop can be enabled by default.