Server mode: only read $stdin when -s or --stdin argument provided
When using `--server` in a non-tty environment, rubocop hangs reading
STDIN. Avoid this by only reading STDIN when the `-s` or `--stdin`
arguments are provided, in which case text on STDIN is expected.

This change allows `rubocop --server` to be used in non-tty contexts,
such as File Watchers or External Tools in RubyMine.