Fix offences in calls inside blocks with braces for Style/MethodCallWithArgsParentheses
For the `omit_parentheses` enforced style, I noticed that calls with
parentheses in calls **inside** braced blocks were not linted in a code
base of mine, so I investigated. It turned out I had a bad detection of
the callers that the braced blocks are attached to, they do **need** the
parentheses for the call to be syntactically correct. However, the
detection leaked into the calls inside the braces that **don't need**
the parentheses.

Example:

```ruby
client.images.map { Image.new _1 }
```

```lisp
(numblock
  (send
    (send
      (send nil :client) :images) :map) 1
  (send
    (const nil :Image) :new
    (lvar :_1)))
```

Here, `(send (const nil :Image) :new (lvar :_1)))`, was considered a
**call with braced block**, because the detection used to check for the
parent node being a block node. This is not correct. I have used the
mechanism defined in `RuboCop::AST::MethodDispatchNode#block_mode` to
properly detect only the send node the block is attached to and not
other sends in the body.