[Fix #11908] Support `AllowedReceivers` for `Style/CollectionMethods`
Fixes #11908.

This PR supports `AllowedReceivers` for `Style/CollectionMethods`.
For example, by setting `AllowedReceivers: ['params']` in RuboCop Rails or user configuration,
the following false positive for `params` can be prevented:

```ruby
def selection_params(current, last_category = nil, params = {})
  params.merge(
    category_id: current.is_a?(Category) ? current.to_param : last_category.to_param,
    site_id: current.is_a?(Site) ? current.to_param : nil,
  ).delete_if { |_k, v| v.nil? }
end
```

The configuration to `params` is not set in the RuboCop core due to Rails specific issues.
So if this approach is accepted, I will add it to RuboCop Rails's default configuration.