Support safe or unsafe autocorrect config for LSP
This PR supports safe or unsafe autocorrect config for LSP.

The autocorrection is still safe by default (`rubocop -a`).

LSP client can switch to unsafe autocorrection (`rubocop -A`) by passing the following
`safeAutocorrect` parameter in the `initialize` request.

```json
{
  "jsonrpc": "2.0",
  "id": 42,
  "method": "initialize",
  "params": {
    "initializationOptions": {
      "safeAutocorrect": false
    }
  }
}
```

Some users may want to set unsafe autocorrection as the default for the LSP,
understanding the risks involved.

However, considering that the default should be safe, unsafe is an option.