[Doc] Tweak the LSP doc
I've observed an instance where a user manually execute `rubocop --lsp`,
which is unnecessary for starting the LSP server.
This PR will clarify in the documentation that it is not necessary to
run `rubocop --lsp`.

And, the sections have been reorganized in the order of most likely to
draw user interest.