[Fix #12020] Fix an error for `Layout/SpaceAfterComma` with `Layout/SpaceBeforeSemicolon`
Fixes #12020.

This PR fixes an infinite loop error for `Layout/SpaceAfterComma`
with `Layout/SpaceBeforeSemicolon` when autocorrection conflicts.

The use of semicolon after comma (`,;`) is limited, so making it difficult to default
the presence or absence of space between the semicolon and the comma.
This PR aims to prevent an infinite loop error occurrences by always allowing the absence of
space after comma when it is followed by semicolon.