[Docs] Add reference to LSP to quickstart section
This PR adds reference to RuboCop's built-in LSP to quickstart section in the README.

While I recognize that the quickstart section is expanding,
I think that due to its potential to dramatically change the development experience,
LSP is worth mentioning as a second form of magic.

To keep the quickstart text concise, I have not mentioned the required RuboCop version.
It can be easily found by referring to the linked page.